Move config code
